Chore(ProfileManager): rename files and fix imports
