+++ packages/playground/src/components/profile_manager/WalletPassword.vue
